Attept to test also jmodless jdk by running from copy of jdk and removing jmods (#7)
* testingb oth jmod-full and jmod-less jdks

jdk25 can be delivered with, or without jmods and should work in both modes
Although current version is shipped without jmods, the first one was not, and they may reapear.
Added GH, where jmods are removed. The action is run even if jmods are missing (thus duplicating the original one) as I have not foudn way how to skip execution and pretend it is green or yellow.